api/pyproject.toml
[project]
name = "fomina-shop-api"
version = "1.0.0"
description = "API for fomina-shop"
authors = [{ name = "VodoGamer", email = "vodogamer@yandex.ru" }]
dependencies = [
	"fastapi>=0.110.2",
	"pydantic>=2.7.0",
	"uvicorn>=0.29.0",
	"sqlalchemy>=2.0.29",
	"alembic>=1.13.1",
	"asyncpg>=0.29.0",
	"python-dotenv>=1.0.1",
	"sqladmin>=0.16.1",
	"fastapi-storages>=0.3.0",
	"pillow>=10.3.0",
	"itsdangerous>=2.2.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["pyright>=1.1.359", "pre-commit>=3.7.0"]

[tool.ruff]
lint.select = ["I", "E", "F", "B", "Q"]
lint.ignore = ["B008"]
line-length = 99

[tool.pyright]
venvPath = "api/"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api"]
